try(backend): revert to global state
I found the cause of https://github.com/IgnisDa/ryot/pull/232.
It was because the scheduler database having to go into a file
based backend. I switched to memory database for my instance
and now will see if this beta release makes memeory usage
better.

+++ apps/backend/src/config.rs
@@ -365,7 +365,7 @@ pub struct SchedulerConfig {
     /// The number of jobs to process every 5 seconds when updating metadata in
     /// the background.
     #[setting(default = 5)]
-    pub rate_limit_num: u64,
+    pub rate_limit_num: i32,
     /// Deploy a job every x hours that performs user cleanup and summary
     /// calculation.
     #[setting(default = 12)]
